extractor/requirements.txt
pandas
numpy
openmeteo_requests
python-dotenv
pyyaml
requests
google-cloud-storage
